Move reusable code to caliban-client
